cleaned up uri tests
