+++ package.json
@@ -22,7 +22,6 @@
     "@types/node": "^15.6.2",
     "@typescript-eslint/eslint-plugin": "^4.26.0",
     "@typescript-eslint/parser": "^4.26.0",
-    "ava": "^4.3.0",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^28.1.2",
     "esbuild": "^0.14.48",
